% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{orderly-deprecated}
\alias{orderly-deprecated}
\alias{pull_archive}
\alias{pull_dependencies}
\alias{set_default_remote}
\alias{unzip_archive}
\title{Deprecated functions}
\usage{
pull_archive(...)

pull_dependencies(...)

set_default_remote(...)

unzip_archive(...)
}
\arguments{
\item{...}{Arguments that will be passed through to the current
function; see the appropriate help there, and ideally update
your code to use the new function.}
}
\description{
Deprecated functions that will be removed from \code{orderly} in
the near future.  Please update to use the new version.
}
\details{
Use of any of these functions will produce a warning that will
indicate the new version of the function to use.

For \code{pull_archive} use
\code{\link{orderly_pull_archive}}

For \code{pull_dependencies} use
\code{\link{orderly_pull_dependencies}}

For \code{set_default_remote} use
\code{\link{orderly_default_remote_set}}

For \code{unzip_archive} use
\code{\link{orderly_unzip_archive}}
}
\keyword{internal}
