% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{orderly_cleanup}
\alias{orderly_cleanup}
\title{Orderly cleanup}
\usage{
orderly_cleanup(name = NULL, config = NULL, locate = TRUE,
  draft = TRUE, data = TRUE, failed_only = FALSE)
}
\arguments{
\item{name}{Optional name; in this case only clean up drafts with this name}

\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{draft}{Remove drafts?}

\item{data}{Remove dangling data (data not used by any draft or
archived report).}

\item{failed_only}{Delete only failed reports (those without the
end-of-run metadata).  This will also clean up drafts created by
\code{\link{orderly_test_start}}}
}
\description{
Clean up orderly draft and data directories.  Deletes all drafts
(possibly just for a set of report names) and then deletes
dangling data sets that are not pointed to by any draft or
committed reports
}
