% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{orderly_config}
\alias{orderly_config}
\title{Retrieve orderly config object.}
\usage{
orderly_config(root = NULL, locate = TRUE)
}
\arguments{
\item{root}{The path to an orderly root directory, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}.}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}
}
\value{
An R6 object representing the orderly config.
}
\description{
Retrieve orderly config object.
}
\examples{
# The orderly demo, with lots of potential reports:
path <- orderly::orderly_example("demo")

orderly::orderly_config(path)
}
