% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{orderly_connect}
\alias{orderly_connect}
\alias{orderly_db}
\title{Connect to orderly databases}
\usage{
orderly_connect(config = NULL, locate = TRUE)

orderly_db(type, config = NULL, locate = TRUE)
}
\arguments{
\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{type}{The type of connection to make (\code{source},
\code{destination}, \code{csv} or \code{rds}).}
}
\description{
Connect to the orderly databases.  \code{orderly_connect} returns
a list with all four connections (source, destination, csv and
rds) while \code{orderly_db} returns a single connection to one of
these.  These should be treated as as \emph{read only} (with the
exception of \code{source}).
}
