% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{orderly_db}
\alias{orderly_db}
\title{Connect to orderly databases}
\usage{
orderly_db(type, config = NULL, locate = TRUE, validate = TRUE)
}
\arguments{
\item{type}{The type of connection to make (\code{source},
\code{destination}, \code{csv} or \code{rds}).}

\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{validate}{Logical, indicating if the database schema should
be validated on open (currently only applicable with \code{type
= "destination"}).  This is primarily intended for internal use.}
}
\description{
Connect to the orderly databases.  These should be treated as as
\emph{read only} (with the exception of \code{source}).
}
