% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{orderly_dependencies}
\alias{orderly_dependencies}
\title{Get the dependencies or dependents of a set of reports from orderly src}
\usage{
orderly_dependencies(
  reports,
  root = NULL,
  locate = TRUE,
  ref = NULL,
  direction = "downstream"
)
}
\arguments{
\item{reports}{Set of report names to get dependencies or dependents of}

\item{root}{The path to an orderly root directory, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}.}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{ref}{Git ref to get dependencies on, if NULL uses currently checked
out branch}

\item{direction}{Downstream to get dependencies of \code{reports} or upstream
to get reports which depend on \code{reports}.}
}
\value{
A list with names matching \code{reports} and values their
dependencies or dependents
}
\description{
For a set of reports find its immediate dependencies (when
\code{direction = "upstream"}) i.e. the reports which this report depends on
or its immediate dependents (when \code{direction = "downstream"}) i.e. the
reports which use this report.
}
\examples{
path <- orderly::orderly_example("demo")

orderly::orderly_dependencies("use_dependency_2", root = path)
orderly::orderly_dependencies("use_dependency_2", direction = "upstream")
}
