% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{orderly_deploy_shiny}
\alias{orderly_deploy_shiny}
\title{Deploy shiny apps onto server}
\usage{
orderly_deploy_shiny(dest, info = "shiny.yml", config = NULL,
  locate = TRUE)
}
\arguments{
\item{dest}{Destination (path) for the shiny apps}

\item{info}{Name of the shiny metadata file (relative to orderly root)}

\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}
}
\description{
Deploy generated shiny apps onto a shiny server
}
