% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{orderly_graph_out_of_date}
\alias{orderly_graph_out_of_date}
\title{Given a tree return a list of reports to be re-run (and the order
that they should be re-run)}
\usage{
orderly_graph_out_of_date(tree)
}
\arguments{
\item{tree}{A dependency tree object from orderly_graph_out_of_date}
}
\value{
a list of report names to be re-run. First report to rerun first
}
\description{
Given a tree return a list of reports to be re-run (and the order
that they should be re-run)
}
