% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{orderly_list}
\alias{orderly_list}
\title{List orderly reports}
\usage{
orderly_list(path = NULL, locate = TRUE)
}
\arguments{
\item{path}{The path to an orderly root directoy, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}
}
\description{
List the \emph{names} of reports known to orderly.  These are the
\emph{source} names, not the results of running reports.
}
