% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{orderly_log_start}
\alias{orderly_log_start}
\alias{orderly_log_stop}
\alias{orderly_log}
\title{Start and stop log}
\usage{
orderly_log_start()

orderly_log_stop()

orderly_log(topic, value)
}
\arguments{
\item{topic}{Up to 9 character text string with the log topic}

\item{value}{Character string with the log entry}
}
\value{
\code{orderly_log_start} invisibly returns a logical
  indicating if logging was previously enabled.  This allows
  patterns like:
\preformatted{if (!orderly::orderly_log_start()) {
  orderly::orderly_log_stop()
}
}
to have a scoped log (i.e., log for the duration of a function).
}
\description{
Start and stop the orderly log.  Soon this might swap out for
\code{loggr}, but for now this should do.  When active, some
actions will print diagnostic information to the message stream.

Send an entry to the orderly log.  This is designed primarily for
use with pacakges that build off of orderly, so that they can log
in a consistent way.
}
\details{
The interface here will change by adding arguments.  Future versions
may support logging to a file.
}
