% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{orderly_out_of_date_reports}
\alias{orderly_out_of_date_reports}
\title{Given a tree return a list of reports to be re-run (and the order
that they should be re-run)}
\usage{
orderly_out_of_date_reports(tree)
}
\arguments{
\item{tree}{A dpendency tree object from orderly_dependency_tree}
}
\value{
a list of report names to be re-run. First report to rerun first
}
\description{
Given a tree return a list of reports to be re-run (and the order
that they should be re-run)
}
