% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{orderly_progress}
\alias{orderly_progress}
\title{Signal a progress message which can be picked up by orderly server}
\usage{
orderly_progress(value)
}
\arguments{
\item{value}{The value of the progress message, can be arbitrary data}
}
\value{
Signals a condition, called for side effect
}
\description{
Use to send messages back to server about progress of a particular
report run.
}
\keyword{internal}
