% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_path.R
\name{orderly_remote_path}
\alias{orderly_remote_path}
\title{Orderly remote at a different path}
\usage{
orderly_remote_path(path, name = NULL)
}
\arguments{
\item{path}{Path to the orderly store}

\item{name}{Name of the remote}
}
\description{
Orderly remote at a different path
}
