% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{orderly_use_resource}
\alias{orderly_use_resource}
\title{Add a resource to orderly.yml}
\usage{
orderly_use_resource(
  resources,
  name = NULL,
  root = NULL,
  locate = FALSE,
  show = TRUE,
  edit = TRUE,
  prompt = TRUE
)
}
\arguments{
\item{resources}{Character vector of resources to add.  These must
be filenames relative to the report directory, must exist, and
must not already be present in the orderly.yml}

\item{name}{Name of the report to modify.  Like
\code{\link{orderly_develop_start}} this can be \code{NULL} if
you have already set the working directory to be the source
directory.}

\item{root}{The path to an orderly root directory, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}.}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{show}{Logical, indicating if we should print the proposed
changes to screen}

\item{edit}{Logical, indicating if we should actually edit the
\code{orderly.yml} file.}

\item{prompt}{Logical, indicating if we should prompt before
editing the orderly.yml file.  Only has an effect if \code{edit}
is \code{TRUE}.}
}
\value{
Invisibly, this function returns information about the
  file it would edit.  This information is primarily for debugging
  purposes and the format is subject to change.
}
\description{
Add one or more resources to an \code{orderly.yml} file.
}
