% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{push_archive}
\alias{push_archive}
\title{Push an archive report to a remote location}
\usage{
push_archive(name, id = "latest", config = NULL, locate = TRUE,
  remote = NULL)
}
\arguments{
\item{name}{Name of the report to download dependencies for}

\item{id}{The identifier (for \code{pull_archive}.  The default is
to use the latest report.}

\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{remote}{Description of the location.  This must be an
\code{orderly_remote} object.}
}
\description{
Push an archive report to a remote location.
}
\details{
This is experimental and only supported using paths for
\code{remote}.  It will be useful for doing something like sharing
preliminary artefacts peer-to-peer before running centrally.  It
is not supported (yet) for pushing onto a montagu server, but that
would be nice to have (probably locked down the same way that the
\code{ref} argument is).
}
