% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{unzip_archive}
\alias{unzip_archive}
\title{Unzip orderly archive into place}
\usage{
unzip_archive(zip, root, name, id)
}
\arguments{
\item{zip}{Path to a zip file}

\item{root}{Orderly root}

\item{name}{Name of the expected report}

\item{id}{Id of the expected report}
}
\description{
Unzip an orderly archive into place.  This exists to work around
an alarming number of failure modes that
\code{\link{utils::unzip}} can suffer.  This should rarely be
needed to be used from user code and exists to support remotes.
}
