% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{set_default_remote}
\alias{set_default_remote}
\alias{get_default_remote}
\title{Set default remote location}
\usage{
set_default_remote(value, root = NULL, locate = TRUE)

get_default_remote(root = NULL, locate = TRUE)
}
\arguments{
\item{value}{A string describing a remote, or \code{NULL} to clear}

\item{root}{The path to an orderly root directoy, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}
}
\description{
Set and get default remote locations
}
