% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{set_default_remote}
\alias{set_default_remote}
\alias{get_default_remote}
\title{Set default remote location}
\usage{
set_default_remote(value)

get_default_remote(config)
}
\arguments{
\item{value}{An \code{orderly_remote_location} object.}
}
\description{
Set and get default remote locations
}
