% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_commit.R
\name{orderly_commit}
\alias{orderly_commit}
\title{Commit a generated report}
\usage{
orderly_commit(id, name = NULL, config = NULL, locate = TRUE)
}
\arguments{
\item{id}{The identifier of the report}

\item{name}{The name of the report - this can be ommited and the
name will be determined from the \code{id}.}

\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}
}
\description{
Commit a generated report
}
