% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{orderly_example}
\alias{orderly_example}
\title{Set up an orderly example}
\usage{
orderly_example(name, path = tempfile())
}
\arguments{
\item{name}{Name of the example}

\item{path}{Destination to create the example - if it exists
already it must be an empty directory.  By default, creates a
new temporary directory}
}
\value{
Returns the path to the orderly example
}
\description{
Set up one of the orderly examples included with the package.
These are not intended to be starting points for new orderly
repositories, but are used in the package examples and vignettes.
}
\examples{
# Create a new copy of the "minimal" example
path <- orderly::orderly_example("minimal")
dir(path)

# Example reports within this repository:
orderly::orderly_list(path)
}
