% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{orderly_init}
\alias{orderly_init}
\title{Initialise an orderly store}
\usage{
orderly_init(root, doc = TRUE, quiet = FALSE)
}
\arguments{
\item{root}{The root of the store; this must be an empty directory
or the path of a directory to create}

\item{doc}{Logical, indicating if documentation should be added to
the directories.  This also has the (potentially useful) effect
of making these directories noticeable by git.}

\item{quiet}{Logical, indicating if informational messages should
be suppressed.}
}
\description{
Initialise an orderly store
}
