% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{orderly_latest}
\alias{orderly_latest}
\title{Find most recent report}
\usage{
orderly_latest(name = NULL, path = NULL, locate = TRUE,
  draft = FALSE, must_work = TRUE)
}
\arguments{
\item{name}{Name of the report to find; if \code{NULL} returns the
most recent report across all names}

\item{path}{The path to an orderly root directoy, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{draft}{Find most recent \emph{draft} report}

\item{must_work}{Throw an error if no report is found.  If FALSE,
returns \code{NA_character_}.}
}
\description{
Find most recent version of an orderly report
}
