% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{orderly_list_drafts}
\alias{orderly_list_drafts}
\alias{orderly_list_archive}
\title{List draft and archived reports}
\usage{
orderly_list_drafts(config = NULL, locate = TRUE)

orderly_list_archive(config = NULL, locate = TRUE)
}
\arguments{
\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}
}
\description{
List draft and archived reports.  This returns a data.frame with
columns \code{name} (see \code{\link{orderly_list}} and \code{id}.
It will expand in future.
}
