% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{orderly_log_on}
\alias{orderly_log_on}
\alias{orderly_log_off}
\alias{orderly_log}
\title{Start and stop log}
\usage{
orderly_log_on()

orderly_log_off()

orderly_log(topic, value)
}
\arguments{
\item{topic}{Up to 9 character text string with the log topic}

\item{value}{Character string with the log entry}
}
\value{
\code{orderly_log_on} and \code{orderly_log_off} invisibly
  returns a logical indicating if logging was previously enabled.
  This allows patterns like:

\preformatted{if (!orderly::orderly_log_off()) {
  orderly::orderly_log_on()
}
}
to disable logging within a function
}
\description{
Start and stop the orderly log.  When active, some actions will
print diagnostic information to the message stream.  This is set
to be on by default.

Send an entry to the orderly log.  This is designed primarily for
use with packages that build off of orderly, so that they can log
in a consistent way.
}
\details{
The interface here will change by adding arguments.  Future versions
may support logging to a file.
}
