% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{orderly_new}
\alias{orderly_new}
\title{Create new report}
\usage{
orderly_new(name, config = NULL, locate = FALSE, quiet = FALSE)
}
\arguments{
\item{name}{Name of the new report (will be a directory name).}

\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{quiet}{Logical, indicating if informational messages should
be suppressed.}
}
\description{
Create new report
}
