% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{orderly_open}
\alias{orderly_open}
\title{}
\usage{
orderly_open(id, name = NULL, config = NULL, locate = TRUE,
  draft = NULL)
}
\arguments{
\item{id}{The identifier of the report - can be \code{latest}, in
which case \code{name} and \code{draft} must be specified}

\item{name}{The name of the report.  Can be omitted if \code{id}
is not \code{latest}}

\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{draft}{Logical, indicating if a draft report should be
found.  Practically only useful when \code{id = "latest"} but
might be useful to ensure presence of a particular type of
report.}
}
\description{
Open the directory for a completed orderly report
}
\author{
Rich FitzJohn
}
