% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{orderly_publish_remote}
\alias{orderly_publish_remote}
\title{Publish orderly report on remote server}
\usage{
orderly_publish_remote(name, id, value = TRUE, remote = NULL)
}
\arguments{
\item{name}{Name of the report (unlike
\code{\link{orderly_publish}} this must be provided sorry)}

\item{id}{The report id}

\item{value}{As \code{\link{orderly_publish}}, \code{TRUE} or
\code{FALSE} to publish or unpublish a report (respectively)}

\item{remote}{Description of the location.  This must be an
\code{orderly_remote} object.}
}
\description{
Publish a report on a remote server
}
