% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{orderly_rebuild}
\alias{orderly_rebuild}
\title{Rebuild the report database}
\usage{
orderly_rebuild(config = NULL, locate = TRUE, verbose = TRUE)
}
\arguments{
\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{verbose}{Logical, indicating if information about the
rebuild should be printed as it runs}
}
\description{
Rebuild the report database
}
