% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{orderly_rebuild}
\alias{orderly_rebuild}
\title{Rebuild the report database}
\usage{
orderly_rebuild(path = NULL, locate = TRUE, verbose = TRUE,
  if_schema_changed = FALSE)
}
\arguments{
\item{path}{The path to an orderly root directoy, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{verbose}{Logical, indicating if information about the
rebuild should be printed as it runs}

\item{if_schema_changed}{Logical, indicating if the rebuild should
take place only if the schema has changed.  This is designed to
be safe to use in (say) deployment scripts because it will be
fast enough to call regularly.}
}
\description{
Rebuild the report database
}
