% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{orderly_run_remote}
\alias{orderly_run_remote}
\title{Run a report on montagu}
\usage{
orderly_run_remote(name, parameters = NULL, ref = NULL, timeout = 3600,
  poll = 1, open = TRUE, stop_on_error = TRUE, progress = TRUE,
  remote = NULL)
}
\arguments{
\item{name}{Name of the report}

\item{parameters}{Parameters for the reprt}

\item{ref}{Optional reference, indicating which branch should be
used.  This cannot be used on production.}

\item{timeout}{Time to wait for the report to be returned (in seconds)}

\item{poll}{Period to poll the server for results (in seconds)}

\item{open}{Logical, indicating if the report should be opened in
a browser on completion}

\item{stop_on_error}{Logical, indicating if we should throw an
error if the report fails.  If you set this to \code{FALSE} it
will be much easier to debug, but more annoying in scripts.}

\item{progress}{Logical, indicating if a progress spinner should
be included.}

\item{remote}{Description of the location.  This must be an
\code{orderly_remote} object.}
}
\description{
Run a report on a remote server (for now this means montagu).  Be
careful doing this because once a report is run to completion on
production it cannot be deleted.  So get things working locally,
test on science and then run on production.
}
