% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{orderly_run_remote}
\alias{orderly_run_remote}
\title{Run a report on a remote server}
\usage{
orderly_run_remote(name, parameters = NULL, ref = NULL,
  timeout = 3600, poll = 1, open = TRUE, stop_on_error = TRUE,
  progress = TRUE, config = NULL, locate = TRUE, remote = NULL)
}
\arguments{
\item{name}{Name of the report}

\item{parameters}{Parameters for the reprt}

\item{ref}{Optional reference, indicating which branch should be
used.  This cannot be used if the remote has \code{master_only}
set.}

\item{timeout}{Time to wait for the report to be returned (in seconds)}

\item{poll}{Period to poll the server for results (in seconds)}

\item{open}{Logical, indicating if the report should be opened in
a browser on completion}

\item{stop_on_error}{Logical, indicating if we should throw an
error if the report fails.  If you set this to \code{FALSE} it
will be much easier to debug, but more annoying in scripts.}

\item{progress}{Logical, indicating if a progress spinner should
be included.}

\item{config}{An orderly configuration, or the path to one (or
\code{NULL} to locate one if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{remote}{Description of the location.  Typically this is a
character string indicating a remote specified in the
\code{remotes} block of your \code{orderly_config.yml}.  It is
also possible to pass in a directly created remote object (e.g.,
using \code{\link{orderly_remote_path}}, or one provided by
another package).  If left \code{NULL}, then the default remote
for this orderly repository is used - by default that is the
first listed remote.}
}
\description{
Run a report on a remote server.
}
