% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{orderly_runner}
\alias{orderly_runner}
\title{Orderly runner}
\usage{
orderly_runner(path, allow_ref = TRUE)
}
\arguments{
\item{path}{Path to use}

\item{allow_ref}{Allow git to change branches/ref for run}
}
\description{
An orderly runner.  This is used to run reports remotely.  It's
designed to be used in conjunction with montagu-reporting-api, so
there is no "draft" stage.
}
