% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{orderly_runner}
\alias{orderly_runner}
\title{Orderly runner}
\usage{
orderly_runner(path, timeout = 2, max_processes = 4L)
}
\arguments{
\item{path}{Path to use}

\item{timeout}{Timeout to use while waiting for R subprocess to
start on run}

\item{max_processes}{The maximum number of concurrent R processes to run.}
}
\description{
An orderly runner.
}
