% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{orderly_workflow}
\alias{orderly_workflow}
\title{Run a workflow.}
\usage{
orderly_workflow(name, root = NULL, locate = TRUE)
}
\arguments{
\item{name}{Name of workflow to run}

\item{root}{The path to an orderly root directory, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}.}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}
}
\value{
ID of workflow run
}
\description{
WIP - skeleton implementation
}
\examples{
path <- orderly::orderly_example("demo")

# To run most reports, provide the report name (and the path if
# not running in the working directory, as is the case here):
#id <- orderly::orderly_workflow("my_workflow", root = path)
}
