% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{pull_dependencies}
\alias{pull_dependencies}
\title{Download dependent reports}
\usage{
pull_dependencies(name, config = NULL, locate = TRUE)
}
\arguments{
\item{name}{Name of the report to download dependencies for}
}
\description{
Download dependent reports from Montagu.  This requires the
montagu package and for montagu's credentials to be correctly set
up.
}
