% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{pull_dependencies}
\alias{pull_dependencies}
\alias{pull_archive}
\title{Download dependent reports}
\usage{
pull_dependencies(name, path = NULL, locate = TRUE, remote = NULL)

pull_archive(name, id = "latest", path = NULL, locate = TRUE,
  remote = NULL)
}
\arguments{
\item{name}{Name of the report to download dependencies for}

\item{path}{The path to an orderly root directoy, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{remote}{Description of the location.  Typically this is a
character string indicating a remote specified in the
\code{remotes} block of your \code{orderly_config.yml}.  It is
also possible to pass in a directly created remote object (e.g.,
using \code{\link{orderly_remote_path}}, or one provided by
another package).  If left \code{NULL}, then the default remote
for this orderly repository is used - by default that is the
first listed remote.}

\item{id}{The identifier (for \code{pull_archive}.  The default is
to use the latest report.}
}
\description{
Download dependent reports.
}
\details{
The \code{pull_archive} function pulls report directly (without it
being a dependent report).

After setting your username up you can run
\code{pull_dependencies("reportname")} to pull the
\emph{dependencies} of \code{"reportname"} down so that
\code{"reportname"} can be run, or you can run
\code{pull_archive("reportname")} to pull a copy of
\code{"reportname"} that has been run on the remote server.
}
