% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{push_archive}
\alias{push_archive}
\title{Push an archive report to a remote location}
\usage{
push_archive(name, id = "latest", root = NULL, locate = TRUE,
  remote = NULL)
}
\arguments{
\item{name}{Name of the report to download dependencies for}

\item{id}{The identifier (for \code{pull_archive}.  The default is
to use the latest report.}

\item{root}{The path to an orderly root directoy, or \code{NULL}
(the default) to search for one from the current working
directory if \code{locate} is \code{TRUE}).}

\item{locate}{Logical, indicating if the configuration should be
searched for.  If \code{TRUE} and \code{config} is not given,
then orderly looks in the working directory and up through its
parents until it finds an \code{orderly_config.yml} file.}

\item{remote}{Description of the location.  Typically this is a
character string indicating a remote specified in the
\code{remotes} block of your \code{orderly_config.yml}.  It is
also possible to pass in a directly created remote object (e.g.,
using \code{\link{orderly_remote_path}}, or one provided by
another package).  If left \code{NULL}, then the default remote
for this orderly repository is used - by default that is the
first listed remote.}
}
\description{
Push an archive report to a remote location.
}
\details{
This is experimental and only supported using
\code{\link{orderly_remote_path}} remotes.  It might be useful for
doing something like sharing preliminary artefacts peer-to-peer
before running centrally.
}
